pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}

[project]
name = "TFM-Labos"
version = "0.1.0"
dependencies = ["rich", "typer", "InquirerPy", "pyyaml"]